# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project-R (DeepQuant) API",
    "version": "2.0.0",
    "description": "AI-driven algorithmic trading platform with R-Factor market intelligence, Historify data management, quantitative backtesting, and AI-powered trade decisions. Built on Dhan V2 for Indian equity & F&O markets.",
    "contact": {
      "name": "Project-R"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/api/r-factor-history": {
      "get": {
        "tags": ["R-Factor"],
        "summary": "R-Factor history and daily leaderboard",
        "description": "Three modes: (1) `?symbol=X` — per-symbol R-Factor trend over N days with ADX, Z-scores, raw bhavcopy. (2) `?date=YYYY-MM-DD` — daily leaderboard ranked by R-Factor. (3) `?dates=true` — list available bhavcopy dates.",
        "operationId": "getRFactorHistory",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "Stock symbol for history mode (e.g., RELIANCE)",
            "schema": { "type": "string" }
          },
          {
            "name": "days",
            "in": "query",
            "description": "Number of days of history (symbol mode)",
            "schema": { "type": "integer", "default": 25 }
          },
          {
            "name": "date",
            "in": "query",
            "description": "Date for daily leaderboard (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max stocks in leaderboard",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "name": "dates",
            "in": "query",
            "description": "Set to `true` to get list of available dates",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "History data returned",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/RFactorHistoryResponse" },
                    { "$ref": "#/components/schemas/RFactorLeaderboardResponse" },
                    { "$ref": "#/components/schemas/AvailableDatesResponse" }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameter",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Insufficient data for symbol or no data for date",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/api/historify/export": {
      "get": {
        "tags": ["Historify"],
        "summary": "Export data as CSV",
        "description": "Export historical candle data as CSV file download. Columns: Symbol, Exchange, Interval, Timestamp, DateTime_IST, Open, High, Low, Close, Volume.",
        "operationId": "exportData",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "description": "Comma-separated symbols",
            "schema": { "type": "string" }
          },
          {
            "name": "interval",
            "in": "query",
            "description": "Candle interval",
            "schema": { "type": "string", "enum": ["Daily", "5min", "15min", "30min", "60min", "1min"] }
          },
          {
            "name": "preset",
            "in": "query",
            "description": "Date range preset",
            "schema": {
              "type": "string",
              "enum": ["Last 7 Days", "Last 30 Days", "Last 90 Days", "Year to Date", "Last 1 Year", "All Time"]
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": { "type": "string", "enum": ["csv"], "default": "csv" }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file download",
            "content": {
              "text/csv": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/api/historify/watchlist": {
      "get": {
        "tags": ["Historify"],
        "summary": "Get watchlist",
        "operationId": "getWatchlist",
        "responses": {
          "200": {
            "description": "Watchlist items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/WatchlistItem" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Historify"],
        "summary": "Add to watchlist",
        "operationId": "addToWatchlist",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["symbol"],
                "properties": {
                  "symbol": { "type": "string" },
                  "exchange": { "type": "string", "default": "NSE" },
                  "segment": { "type": "string" },
                  "securityId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated watchlist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/WatchlistItem" }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Historify"],
        "summary": "Remove from watchlist",
        "operationId": "removeFromWatchlist",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["symbol"],
                "properties": {
                  "symbol": { "type": "string" },
                  "exchange": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated watchlist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/WatchlistItem" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ZScores": {
        "type": "object",
        "description": "Z-score normalized values for all 8 factors.",
        "properties": {
          "spread": { "type": "number", "description": "(H-L)/close ratio vs 20d avg — dominant predictor" },
          "oi_level": { "type": "number", "description": "OI ratio vs 20d avg (>1 = accumulation, <1 = distribution)" },
          "fut_turnover": { "type": "number", "description": "Futures turnover Z-score" },
          "fut_volume": { "type": "number", "description": "Futures volume Z-score" },
          "opt_volume": { "type": "number", "description": "Options total volume Z-score" },
          "eq_trade_size": { "type": "number", "description": "Equity avg trade size Z-score" },
          "oi_change": { "type": "number", "description": "|OI change| Z-score" },
          "pcr": { "type": "number", "description": "Put-Call ratio Z-score" }
        }
      },
      "MarketRegime": {
        "type": "string",
        "enum": ["Elephant", "Cheetah", "Hybrid", "Defensive"],
        "description": "Market regime classification based on R-Factor score"
      },
      "ModelType": {
        "type": "string",
        "enum": ["ols", "spread-quad", "momentum", "ensemble"],
        "description": "Which model computed the R-Factor"
      },
...
      "WatchlistItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "symbol": { "type": "string" },
          "exchange": { "type": "string", "default": "NSE" },
          "segment": { "type": "string", "default": "EQ" },
          "securityId": { "type": "string", "nullable": true },
          "addedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "RFactorHistoryResponse": {
        "type": "object",
        "title": "Symbol history response",
        "properties": {
          "success": { "type": "boolean" },
          "symbol": { "type": "string" },
          "sector": { "type": "string", "nullable": true },
          "summary": {
            "type": "object",
            "properties": {
              "avgR": { "type": "number" },
              "maxR": { "type": "number" },
              "minR": { "type": "number" },
              "blastDays": { "type": "integer" },
              "totalDays": { "type": "integer" },
              "dominantRegime": { "type": "string" },
              "regimeCounts": { "type": "object", "additionalProperties": { "type": "integer" } },
              "trendDirection": { "type": "string", "enum": ["up", "down", "flat"] },
              "avgSpread": { "type": "number" },
              "maxSpread": { "type": "number" }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string" },
                "compositeRFactor": { "type": "number" },
                "rawRFactor": { "type": "number" },
                "scaledRFactor": { "type": "number" },
                "confidence": { "type": "number" },
                "pctChange": { "type": "number", "nullable": true },
                "delta": { "type": "number", "nullable": true },
                "zScores": { "$ref": "#/components/schemas/ZScores" },
                "regime": { "$ref": "#/components/schemas/MarketRegime" },
                "isBlastTrade": { "type": "boolean" },
                "modelUsed": { "$ref": "#/components/schemas/ModelType" },
                "adx": { "type": "number", "nullable": true },
                "plusDI": { "type": "number", "nullable": true },
                "minusDI": { "type": "number", "nullable": true },
                "raw": {
                  "type": "object",
                  "properties": {
                    "eqHigh": { "type": "number" },
                    "eqLow": { "type": "number" },
                    "eqClose": { "type": "number" },
                    "eqVolume": { "type": "number" },
                    "futVolume": { "type": "number" },
                    "futOi": { "type": "number" },
                    "futOiChange": { "type": "number" },
                    "futTurnover": { "type": "number" },
                    "ceVolume": { "type": "number" },
                    "peVolume": { "type": "number" },
                    "optOi": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      },
      "RFactorLeaderboardResponse": {
        "type": "object",
        "title": "Daily leaderboard response",
        "properties": {
          "success": { "type": "boolean" },
          "date": { "type": "string" },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "symbol": { "type": "string" },
                "sector": { "type": "string", "nullable": true },
                "compositeRFactor": { "type": "number" },
                "rawRFactor": { "type": "number" },
                "scaledRFactor": { "type": "number" },
                "confidence": { "type": "number" },
                "zScores": { "$ref": "#/components/schemas/ZScores" },
                "regime": { "$ref": "#/components/schemas/MarketRegime" },
                "isBlastTrade": { "type": "boolean" },
                "modelUsed": { "$ref": "#/components/schemas/ModelType" }
              }
            }
          }
        }
      },
      "AvailableDatesResponse": {
        "type": "object",
        "title": "Available dates response",
        "properties": {
          "success": { "type": "boolean" },
          "dates": { "type": "array", "items": { "type": "string" } }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        }
      },
...
    }
  }
}